backend/requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
python-dotenv>=1.0.0
pymongo>=4.6.0
certifi>=2023.0.0
bcrypt>=4.0.0
PyJWT>=2.8.0
cloudinary>=1.36.0
opencv-contrib-python>=4.8.0
pillow>=10.0.0
numpy>=1.24.0
pandas>=2.0.0
